Changed N to be long. Added possibility to configure the start date for datetime generators.

+++ SQLDataProducer.Entities/DatabaseEntities/ColumnEntity.cs
@@ -247,7 +247,7 @@ namespace SQLDataProducer.DatabaseEntities.Entities
             writer.WriteEndElement();
         }
 
-        public void GenerateValue(int n)
+        public void GenerateValue(long n)
         {
             if (Generator.GEN_ValueFromOtherColumn == Generator.GeneratorName)
                 return;
